Match note titles in task search and share breadcrumb grouping via core
Cherry-picked from the competing #686: task search now also matches the
source note's title, and groupTaskContexts/visibleTaskBreadcrumbs move
from a desktop-local module into @reflect/core's group-tasks so any
surface can reuse them.
