+++ apps/desktop/src/components/tasks/task-group-section.tsx
@@ -1,8 +1,7 @@
 import { Fragment, type MutableRefObject, type ReactElement } from 'react'
 import { Plus } from 'lucide-react'
-import type { OpenTask, TaskGroup } from '@reflect/core'
+import { groupTaskContexts, type OpenTask, type TaskGroup } from '@reflect/core'
 import { addTargetForGroup, taskGroupHeaderStyle } from '@/lib/tasks/task-group-presentation'
-import { groupTaskContexts } from '@/lib/tasks/task-breadcrumbs'
 import type { InsertTaskTarget } from '@/lib/tasks/task-insert-target'
 import { taskKey } from '@/lib/tasks/task-identity'
 import type { TaskSelection } from '@/lib/tasks/use-task-selection'
