+++ packages/core/src/exports/sync-markdown-indexing.ts
@@ -158,8 +158,10 @@ export {
   getNotesByTag,
   getOpenTasks,
   getCompletedTasks,
+  groupTaskContexts,
   groupTasks,
   taskDateBucket,
+  visibleTaskBreadcrumbs,
   listDailyNotes,
   listNotes,
   listNoteTags,
@@ -196,6 +198,7 @@ export {
   type DuplicateIdGroup,
   type NoteRow,
   type OpenTask,
+  type TaskContext,
   type TaskGroup,
   type TaskGroupKind,
   type NoteListEntry,
